requirements.txt
dice
discord.py
python-dotenv
sqlalchemy
psycopg2
sqlalchemy-utils